requirements.txt
streamlit
openai
PyMuPDF
tqdm 
langchain
langchain-community
langchain-groq
pymupdf
faiss-cpu
sentence-transformers
python-dotenv
pydantic
fpdf
pyperclip